build: Update release-patch to handle bugfixes
Bugfixes are done on branches that don't have an unreleased
section in the CHANGELOG which would cause the last line
to be deleted and the change put at the bottom of the
CHANGELOG. Now there is a test for the unreleased section
which if it exists does the old behavior. If not the new
message will be placed above the last version section.

Fixes: #2533

Signed-off-by: Joshua Shanks <jjshanks@gmail.com>