topdown: respect shallow inlining for functions depending on unknowns
When shallow inlining is configured, we'll now save the PE'ed function
into support and call it if both of these are met:

a) there are unknowns in either the function definitions or the call site
   arguments
b) there is more than one function definition

Since if (a) holds but (b) doesn't, we're still avoiding undesired cross
products, even when inlining the function.

Fixes #3463.

Also:

- topdown/eval: refactor savePackagePathAndTerm

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>