ast: Treat array.items as dynamic element type
This commit fixes the conversion to treat array.items schema as the
dynamic element type in arrays as opposed to the static element
type. If array.items is defined as an object then it applies to ALL
elements in the array. On the other hand, if array.items is defined as
an array then it applies pairwise to the elements in the array.

In order to fix this, we have to expose a new field from the schema library.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>