server: Refactor index.html to use fetch()
This commit refactors the serve to avoid using simple HTML forms to
execute queries against the server. This avoids having the GET /
handler run queries. With this change, the index.html page is now
static (except for templating that happens to inject the OPA version
info on the page.)

This commit drops support for the "explain" option on the form. This
feature is not very useful in it's current state so it should not be missed.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>