topdown: Fix http send flaky test
The affected test case checks caching behavior by
making http requests to the test server. It expects
the test server response body to be the same. However
the Date header could be different between calls.

The test case uses the unification operator to assign a
variable to the result of http.send. Then the next expression
in the rule body performs a comparison which would intermittently
fail due to the changing value of the Date header in the
http.send response.

This change updates the test to use diffferent variables to asssign
the result of http.send and thus avoid the comparison.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>