wasm/sdk: handle random seed needed by 'uuid.rfc4122()'
There random seed to be used with the host-provided builtin uuid.rfc4122
was missing from the BuiltinContext that's created in the dispatcher. It
hadn't been thread through from rego to the vm call.

The unit tests for the builtin now run against both targets.

Also some style adjustments to the builtin code for uuid.rfc4122.

Fixes #3622.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>