ast/compiler: treat local vars introduced by `some` differently (#3581)
Before, we had `some x` introduce a new (generated) local variable.
Generated variables aren't reported as unsafe; so the legitimate
error in

    package p

    r[x] { some x }

would have been ignored, and the evaluation ended in

    var needs evaluation

since a local var made it out of the topdown evaluation process.

Now, we're checking for each unsafe var if it had been rewritten --
for local variables introduced by `some`, that'll yield the user-
provided name (say, `x`), which in turn isn't recognized as
generated.

As similar fix applied to the queryCompiler: Before, running

    opa eval '[1 | some x; x > 1]'

would have noticed that the unsafe variable x was (after rewriting)
a generated variable, and reported that the expression was unsafe:

    1:14: rego_unsafe_var_error: expression is unsafe

Now, the error presented to the user is

    1:14: rego_unsafe_var_error: var x is unsafe

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Dolev Farhi <farhi.dolev@gmail.com>