plugins: Add new Resource configuration for decision logs
As described in #3618, when using a service with a specific path for receiving logs (like a SaaS SIEM) its often the case that the path to the endpoint is static and cannot be changed.
This change allows setting the entire path to be used when sending decision logs via the decision_logs.resource field, unlike the partition configuration which appends the path as /log/<partition>. The uploadChunk function was updated to take both the partition name and resource as parameters, but the partition name will take precendence over resource.

Fixes #3618

Signed-off-by: Casey Buto <cbuto22@gmail.com>