ast: fix panic in ParseModule() with all space chars, starting with tab (#3652)
The specific combination where the first character is a tab, and the rest
are falling under `unicode.IsSpace()` would lead to strings.Repeat being
fed an empty count (to determine the position of the error pointer).

Displaying different "spaces" in a discernable way isn't solved here,
but I think it's also an edge case we can accept as-is.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Dolev Farhi <farhi.dolev@gmail.com>