wasm: put stack first, grow it to 8MiB.
Just want to see the impact -- if any -- of this.

From how I understand this, https://surma.dev/things/c-to-webassembly/,
we should be able to modify the stack size at the time of `opa build`
if we keep the stack after data: it's all a matter of moving
`__heap_base`.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>