wasm: put stack first, adjust heap base (#3660)
Putting the stack first is preferrable for how C/C++'s stack is mapped
in to Wasm: it's assigned a memory pointer, and grows up. Without
putting the stack first, it can grow into the data section and over-
write globals, leading to situations best described as weird.

Putting it first means that if the evaluation runs out of stack space,
a memory-out-of-bounds trap will occur: it'll try to access a negative
memory location.

In internal/compiler/wasm, we append segments to the data section.
Since our memory layout is

|  <-- stack | -- data (llvm, opa) -- | heap -->  |
we need to adjust the border between data and heap, i.e., where the heap
starts. When initializing a module, the Start function emitted by the
compiler will call the opa_malloc_init function with the new heap base.

Also:

* run-wasm-rego-tests.sh: bump node image version
* wasm/graph.unreachable: fix "memory access out of bounds" issue

   We've never seen this bug in the wild before, but due to the memory layout
   change, the second branch -- casting to an array, accessing its fields --
   would now attempt to read something well beyond the end of memory, and trap.

* rego/testdata: remove Makefile and input rego policy

   This test bundle is simple enough to recreate if need be.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Dolev Farhi <farhi.dolev@gmail.com>