website: add contributing section from existing content (#3752)
This is taken as-is from CONTRIBUTING.md for starters. The subsection
markup has been adjusted as there can only be one "# header 1" in a
document. Also done some light copy-editing.

Also adds the DEVELOPMENT.md content, slightly reformatted because of
bugs in our old hugo version. Code blocks in lists would look all wonky.

Both CONTRIBUTING.md and DEVELOPMENT.md now refer to the corresponding
website sections. As of now, it'll be `edge`, we can flip it to `latest`
after the next release.

Fixes #3751.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Dolev Farhi <farhi.dolev@gmail.com>