docs: Make Playground the first link in docs sidemenu
The current location for rego playground is very hard to find. It's almost the last item in a huge list.

It should be one of the most important links. The online playground is how people can _try_ Rego and get instant feedback. We should encourage people to use it as a first port of call when learning.