bundle: Repalce HasPrefix in erasePolicy
Using HasPrefix to compare a root path to the path of policy can result
in unexpected behaviour when different bundles have similarly named
roots.

This replaces the HasPrefix comparison with a comparison that treats the
paths as directories and checks if the policy path is the same path as
the root or a subdirectory of the root path.

Signed-off-by: Edward Paget <edward.paget@chime.com>

Fixed changes requested by @srenatus

Replace custom function with existing one.

Signed-off-by: Edward Paget <edward.paget@chime.com>