website/build: misc changes to website build scripts (#3886)
* website/scripts/load-docs.sh: less verbose output

1. Less output: the tag semver ordering isn't something we work
   on daily, so let's silence that.

2. Use 'git archive' and tar, not a checkout:
   this way, we don't mess up the working directory or reflog. It
   should also be faster, since we don't end up writing all the
   vendor files, binary blobs, etc to disk.

3. Make the 'latest' symlink relative.

* scripts/live-blocks/inject: run preprocess with max allowed args

Before, when attempting to do this, I had used

    find ... -exec npm run preprocess {} +

and find had determined what the maximum number of allowed command
line args were, and executed that call in batches of those. `npm`,
however, would have prepended the call with its own args, thereby
exceeding the limit.

As a way out, we had used `xargs -n200`, to preprocess the files
in batches of 200 at a time.

Now, we're telling `find` the exact command needed, bypassing the
package.json helper definition.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Dolev Farhi <farhi.dolev@gmail.com>