ast/parser/ParseRuleFromExpr: ignore Calls
Consider the following example:

    package p
    in(set, elem) {
      _ = set[elem]
    }
    in({3}, 4)

Up until the latest release (inclusive), the last line would have been
parsed as a shorthand for

    in({3}, 4) = true { true }

With the introduction of the `in` operator, the naked call of line 5 started
raising an error:

    rego_parse_error: rule name conflicts with built-in function

Similar things happen to top-level calls to `print("whatever")`.

Arguably, those extra rule bodies never did what the user intended in the
first place: assert something -- they haven't been inside of a rule body.

With this change, calls like these are simply ignored, and stripped out by
calls to `opa fmt`.

Note that it only applies to calls to existing built-in functions. You are
still able to use shorthand notation for declaring facts, like

    fruit("apple")
    fruit("banana")

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>