ast/parser: fix 'some x in xs' edge case via hinting (#3934)
Encountering the token stream

    some x in [ ... ]

it's hard to tell if this really needs a hint or not, since there are
perfectly valid constructions, e.g. when `in` is a partial rule.

So what we do instead now is to record hints in the parser state, and
discharge them into error messages when an error occurs.

We should not hint less, but it should get us rid of false positives:
if there's no error happening later, the hinting would be ignored.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Dolev Farhi <farhi.dolev@gmail.com>