topdown/eval: special-case 'in' calls to avoid plugging args
The way this was done before, the evaluator had to plug the function
arguments, including the collection, in full, to evaluate them as is
usually done with other functions.

Introducing the special case, and the thunk-y plugNamespaced-variant,
we can avoid that: when evaluating the `x in xs` and `x, y in xs`
calls, the evaluator will plug and compare the collection element
(pairs) one-by-one.

One optimization still left open is that the keys are plugged for
objects, even if the call is for `x in obj` and it could thus be
avoided.

This does not yet affect PE -- we've had saved calls under all
circumstances involving unknown arguments (standard func call behaviour)
-- but it adds a bunch of test cases to set the stage for future
improvements.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>