ast: Fix print call rewriting for calls in head
The print call rewriting only worked for calls in the body of the
rule. If a call was embedded inside of a comprehension in the head of
the rule, it would not get rewritten. This change just updates the
compiler to use WalkBodies instead of WalkRules which ensures that all
ast.Body structures in the module get visited regardless of where they
are contained.

Fixes #3967

Signed-off-by: Torin Sandall <torinsandall@gmail.com>