ast: Fix print call rewriting for calls in head
The print call rewriting only worked for calls in the body of the
rule. If a call was embedded inside of a comprehension in the head of
the rule, it would not get rewritten. This change just updates the
compiler to use WalkBodies in conjunction with WalkRules and ensures
that function args are marked as safe so that cases like the following
do not generate false-positive errors:

	f(x) = {1 | x[y]; print(y)}

Fixes #3967

Signed-off-by: Torin Sandall <torinsandall@gmail.com>