website: bump to hugo to 0.88.1, adapt code and live-blocks (#3986)
* website: bump hugo, updated templates

For some reason, the "and" trick to step around undefined values no longer
worked. So now, we're wrapping {{ if }} conditionals instead: the outer one
checks that the value is defined, the inner one asserts something on its
value.

* live-blocks/preprocess: add wrapper div, remove newline

Somewhere between 0.55 and the most recent version, Hugo stopped emitting
that div on purpose for languages it didn't know.

Since we depend on it for further processing, we're adding it back.

Also, a trailing newline for code elements was introduced, and that broke
our codemirrors integration -- every code entry would be folllowed by an
empty line. Now, we're removing that in the same preprocess step.

* website: fix info|danger blocks
* website/config: trust markdown to include html

Fixes #3787.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>