internal/merge: remove InterfaceMaps benchmark
Due to the fiddling with StartTimer/StopTimer in the inner loop, this
test ended up taking 5 minutes, even though the time per store size was
not that bad.

After playing with righting the wrong, I've concluded that we don't
gain too much from accurate numbers here -- less so than we did when
this change was introduced originally.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>