testcases: remove non-temp error case, it's unreliable
This case worked locally and in all containers I've tried running it
in, but failed with

    eval_builtin_error: net.lookup_ip_addr: lookup :): Temporary failure in name resolution

on github actions. I don't think it's worth figuring this out, since
my intention was never to introduce a test that does an actual network
call. Turns out this test apparently does that -- the weird input is
not caught earlier than that.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>