ast/term: sort set's keys slice on insert
Before, we'd sort a set's `keys` slice in-place, and only when it
was actually compared to another set. This side-effect of a comparison
can be unexpected.

Now, we'll keep the `keys` slice sorted by inserting every new element
into its sorted position.

Analogous to #3823, where we did this with objects' `keys` slices.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>