Fix opa fmt issue on refs containing operators (#4105)
The current version of OPA shows unexpected behavior when formatting files that
contain operators in refs.  This is commonly used to get e.g. the last item out
of an array:

    $ cat test.rego
    package test

    foo = x {
    	arr = [1, 2, 3]
    	x = arr[count(arr) - 1]
    }
    $ opa fmt test.rego
    package test

    foo = x {
    	arr = [1, 2, 3]
    	x = arr[minus(count(arr), 1)]
    }

This fixes that issue by using `writeTerm()` rather than `String()` for the ref.
Another approach could be to change this in `String()`; I wasn't sure which one
was better.

Signed-off-by: Jasper Van der Jeugt <jasper@fugue.co>