opentelemetry: change config, enabled -> type == "grpc"
This is meant to future-proof the config. Future values might include
"http", where the address, TLS and sampling related settings would be
reused; and "global", where the setup would use whatever global trace
provider was configured with otel. The latter use case would be for
embedding OPA via the SDK.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>